+++ sme_ptrf_apps/core/tasks/__init__.py
@@ -13,6 +13,7 @@ from .gerar_notificacao_atraso_entrega_ajustes_pc import gerar_notificacao_atras
 from .gerar_notificacao_proximidade_fim_prazo_ajustes_pc import gerar_notificacao_proximidade_fim_prazo_ajustes_prestacao_conta_async
 from .gerar_xlsx_extrato_dres import gerar_xlsx_extrato_dres_async
 from .gerar_previa_relatorio_acertos import gerar_previa_relatorio_acertos_async
+from .gerar_previa_relacao_apos_acertos import gerar_previa_relatorio_apos_acertos_async
